Moved all server stuff to a separate folder.
Refactored the way the API docs are structured.
TODO: tonic hasn't been upgraded to hyper 1.0 yet. Need to figure out
how to do that
